Move ErrorScreen Handling out of RuntimeInitializeOnLoad
