+++ Assets/Code/src/Runtime/ErrorManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 64825340ec94fed4db243671657386c3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
